+++ cca_zoo/__init__.py
@@ -1,10 +1,11 @@
-import cca_zoo.data
-import cca_zoo.dcca
-import cca_zoo.dccae
-import cca_zoo.deep_models
-import cca_zoo.deepwrapper
-import cca_zoo.dvcca
-import cca_zoo.innerloop
-import cca_zoo.objectives
-import cca_zoo.plot_utils
-import cca_zoo.wrappers
+import data
+import models
+
+try:
+    import deepmodels
+except ModuleNotFoundError:
+    pass
+try:
+    import probablisticmodels
+except ModuleNotFoundError:
+    pass
