+++ cca_zoo/deepmodels/__init__.py
@@ -1,0 +1,9 @@
+import objectives
+from ._dcca_base import _DCCA_base
+from .dcca import DCCA
+from .dcca_noi import DCCA_NOI
+from .dccae import DCCAE
+from .deepwrapper import DeepWrapper
+from .dtcca import DTCCA
+from .dvcca import DVCCA
+from .splitae import SplitAE
