+++ cca_zoo/deepmodels/deepwrapper.py
@@ -6,9 +6,9 @@ import torch
 from torch.utils.data import DataLoader
 from torch.utils.tensorboard import SummaryWriter
 
-import cca_zoo.plot_utils
-from cca_zoo.dcca import _DCCA_base
-from cca_zoo.wrappers import _CCA_Base
+import cca_zoo.utils.plot_utils
+from cca_zoo.deepmodels.dcca import _DCCA_base
+from cca_zoo.models.wrappers import _CCA_Base
 
 
 class DeepWrapper(_CCA_Base):
