+++ cca_zoo/models/__init__.py
@@ -1,0 +1,6 @@
+from .cca_base import _CCA_Base
+from .gcca import GCCA
+from .iterative import _Iterative, PLS, PMD, CCA_ALS, ElasticCCA, ParkhomenkoCCA, SCCA, SCCA_ADMM
+from .mcca import MCCA, KCCA
+from .rcca import CCA, rCCA
+from .tcca import TCCA, KTCCA
