+++ cca_zoo/probablisticmodels/__init__.py
@@ -1,0 +1,1 @@
+from .vcca import VariationalCCA
