+++ example_scripts/example_main.py
@@ -11,19 +11,19 @@ import matplotlib.pyplot as plt
 import numpy as np
 from torch.utils.data import Subset
 
-from cca_zoo import data
-from cca_zoo import wrappers
+from cca_zoo.data import Noisy_MNIST_Dataset
+from cca_zoo.models.wrappers import CCA, SCCA, ElasticCCA, PMD, PLS, KCCA, MCCA, GCCA, TCCA
 
 # Load MNIST Data
 os.chdir('..')
 N = 1000
-dataset = data.Noisy_MNIST_Dataset(mnist_type='MNIST', train=True)
+dataset = Noisy_MNIST_Dataset(mnist_type='MNIST', train=True)
 ids = np.arange(min(2 * N, len(dataset)))
 np.random.shuffle(ids)
 train_ids, val_ids = np.array_split(ids, 2)
 val_dataset = Subset(dataset, val_ids)
 train_dataset = Subset(dataset, train_ids)
-test_dataset = data.Noisy_MNIST_Dataset(mnist_type='MNIST', train=False)
+test_dataset = Noisy_MNIST_Dataset(mnist_type='MNIST', train=False)
 test_ids = np.arange(min(N, len(test_dataset)))
 np.random.shuffle(test_ids)
 test_dataset = Subset(test_dataset, test_ids)
@@ -51,7 +51,7 @@ c1 = [1, 3, 7, 9]
 c2 = [1, 3, 7, 9]
 param_candidates = {'c': list(itertools.product(c1, c2))}
 
-pmd = wrappers.PMD(latent_dims=latent_dims, max_iter=max_iter).gridsearch_fit(
+pmd = PMD(latent_dims=latent_dims, max_iter=max_iter).gridsearch_fit(
     train_view_1,
     train_view_2,
     param_candidates=param_candidates,
@@ -66,7 +66,7 @@ pmd_results = np.stack((pmd.train_correlations[0, 1, :], pmd.predict_corr(test_v
 """
 
 # %%
-linear_cca = wrappers.CCA(latent_dims=latent_dims)
+linear_cca = CCA(latent_dims=latent_dims)
 
 linear_cca.fit(train_view_1, train_view_2)
 
@@ -77,7 +77,7 @@ linear_cca_results = np.stack(
 ### (Regularized) Generalized CCA via alternating least squares (can pass more than 2 views)
 """
 
-gcca = wrappers.GCCA(latent_dims=latent_dims, c=[1, 1])
+gcca = GCCA(latent_dims=latent_dims, c=[1, 1])
 
 gcca.fit(train_view_1, train_view_2)
 
@@ -87,7 +87,7 @@ gcca_results = np.stack((gcca.train_correlations[0, 1], gcca.predict_corr(test_v
 ### (Regularized) Multiset CCA via alternating least squares (can pass more than 2 views)
 """
 
-mcca = wrappers.MCCA(latent_dims=latent_dims, c=[0.5, 0.5])
+mcca = MCCA(latent_dims=latent_dims, c=[0.5, 0.5])
 # small ammount of regularisation added since data is not full rank
 
 mcca.fit(train_view_1, train_view_2)
@@ -98,7 +98,7 @@ mcca_results = np.stack((mcca.train_correlations[0, 1], mcca.predict_corr(test_v
 ### (Regularized) Tensor CCA via alternating least squares (can pass more than 2 views)
 """
 
-tcca = wrappers.TCCA(latent_dims=latent_dims, c=[0.99, 0.99])
+tcca = TCCA(latent_dims=latent_dims, c=[0.99, 0.99])
 # small ammount of regularisation added since data is not full rank
 
 tcca.fit(train_view_1, train_view_2)
@@ -110,7 +110,7 @@ tcca_results = np.stack((tcca.train_correlations[0, 1], tcca.predict_corr(test_v
 """
 
 # %%
-pls = wrappers.PLS(latent_dims=latent_dims)
+pls = PLS(latent_dims=latent_dims)
 
 pls.fit(train_view_1, train_view_2)
 
@@ -126,7 +126,7 @@ c1 = [1, 3, 7, 9]
 c2 = [1, 3, 7, 9]
 param_candidates = {'c': list(itertools.product(c1, c2))}
 
-pmd = wrappers.PMD(latent_dims=latent_dims, max_iter=max_iter).gridsearch_fit(
+pmd = PMD(latent_dims=latent_dims, max_iter=max_iter).gridsearch_fit(
     train_view_1,
     train_view_2,
     param_candidates=param_candidates,
@@ -145,7 +145,7 @@ c1 = [0.00001, 0.0001]
 c2 = [0.00001, 0.0001]
 param_candidates = {'c': list(itertools.product(c1, c2))}
 
-scca = wrappers.SCCA(latent_dims=latent_dims, max_iter=max_iter).gridsearch_fit(
+scca = SCCA(latent_dims=latent_dims, max_iter=max_iter).gridsearch_fit(
     train_view_1,
     train_view_2,
     param_candidates=param_candidates,
@@ -185,14 +185,14 @@ l1_1 = [0.01, 0.1]
 l1_2 = [0.01, 0.1]
 param_candidates = {'c': list(itertools.product(c1, c2)), 'l1_ratio': list(itertools.product(l1_1, l1_2))}
 
-elastic = wrappers.ElasticCCA(latent_dims=latent_dims,
-                              max_iter=max_iter).gridsearch_fit(train_view_1,
-                                                                train_view_2,
-                                                                param_candidates=param_candidates,
-                                                                folds=cv_folds,
-                                                                verbose=True,
-                                                                jobs=jobs,
-                                                                plot=True)
+elastic = ElasticCCA(latent_dims=latent_dims,
+                     max_iter=max_iter).gridsearch_fit(train_view_1,
+                                                       train_view_2,
+                                                       param_candidates=param_candidates,
+                                                       folds=cv_folds,
+                                                       verbose=True,
+                                                       jobs=jobs,
+                                                       plot=True)
 
 elastic_results = np.stack(
     (elastic.train_correlations[0, 1, :], elastic.predict_corr(test_view_1, test_view_2)[0, 1, :]))
@@ -214,11 +214,11 @@ c2 = [0.9, 0.99]
 
 param_candidates = {'kernel': ['linear', 'linear'], 'c': list(itertools.product(c1, c2))}
 
-kernel_reg = wrappers.KCCA(latent_dims=latent_dims).gridsearch_fit(train_view_1, train_view_2,
-                                                                   folds=cv_folds,
-                                                                   param_candidates=param_candidates,
-                                                                   verbose=True, jobs=jobs,
-                                                                   plot=True)
+kernel_reg = KCCA(latent_dims=latent_dims).gridsearch_fit(train_view_1, train_view_2,
+                                                          folds=cv_folds,
+                                                          param_candidates=param_candidates,
+                                                          verbose=True, jobs=jobs,
+                                                          plot=True)
 kernel_reg_results = np.stack((
     kernel_reg.train_correlations[0, 1, :],
     kernel_reg.predict_corr(test_view_1, test_view_2)[0, 1, :]))
@@ -230,11 +230,11 @@ degree2 = [2, 3]
 param_candidates = {'kernel': ['poly', 'poly'], 'degree': list(itertools.product(degree1, degree2)),
                     'c': list(itertools.product(c1, c2))}
 
-kernel_poly = wrappers.KCCA(latent_dims=latent_dims).gridsearch_fit(train_view_1, train_view_2,
-                                                                    folds=cv_folds,
-                                                                    param_candidates=param_candidates,
-                                                                    verbose=True, jobs=jobs,
-                                                                    plot=True)
+kernel_poly = KCCA(latent_dims=latent_dims).gridsearch_fit(train_view_1, train_view_2,
+                                                           folds=cv_folds,
+                                                           param_candidates=param_candidates,
+                                                           verbose=True, jobs=jobs,
+                                                           plot=True)
 
 kernel_poly_results = np.stack((
     kernel_poly.train_correlations[0, 1, :],
@@ -247,11 +247,11 @@ gamma2 = [1e+1, 1e+2, 1e+3]
 param_candidates = {'kernel': ['rbf', 'rbf'], 'sigma': list(itertools.product(gamma1, gamma2)),
                     'c': list(itertools.product(c1, c2))}
 
-kernel_gaussian = wrappers.KCCA(latent_dims=latent_dims).gridsearch_fit(train_view_1, train_view_2,
-                                                                        folds=cv_folds,
-                                                                        param_candidates=param_candidates,
-                                                                        verbose=True, jobs=jobs,
-                                                                        plot=True)
+kernel_gaussian = KCCA(latent_dims=latent_dims).gridsearch_fit(train_view_1, train_view_2,
+                                                               folds=cv_folds,
+                                                               param_candidates=param_candidates,
+                                                               verbose=True, jobs=jobs,
+                                                               plot=True)
 
 kernel_gaussian_results = np.stack((
     kernel_gaussian.train_correlations[0, 1, :],
@@ -267,13 +267,13 @@ We also have deep CCA methods (and autoencoder variants)
 We introduce a Config class from configuration.py. This contains a number of default settings for running DCCA.
 
 """
-from cca_zoo import deepwrapper, objectives, dcca, deep_models
+from cca_zoo.deepmodels import dcca, dccae, dvcca, objectives, architectures, deepwrapper
 
 # %%
 # DCCA
 print('DCCA')
-encoder_1 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
-encoder_2 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
+encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
+encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
 dcca_model = dcca.DCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2])
 
 # hidden_layer_sizes are shown explicitly but these are also the defaults
@@ -285,8 +285,8 @@ dcca_results = np.stack((dcca_model.train_correlations[0, 1], dcca_model.predict
 
 # DGCCA
 print('DGCCA')
-encoder_1 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
-encoder_2 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
+encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
+encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
 dgcca_model = dcca.DCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2], objective=objectives.GCCA)
 
 # hidden_layer_sizes are shown explicitly but these are also the defaults
@@ -299,8 +299,8 @@ dgcca_results = np.stack(
 
 # DMCCA
 print('DMCCA')
-encoder_1 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
-encoder_2 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
+encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
+encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
 dmcca_model = dcca.DCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2], objective=objectives.MCCA)
 
 # hidden_layer_sizes are shown explicitly but these are also the defaults
@@ -313,8 +313,8 @@ dmcca_results = np.stack(
 
 # DCCA_NOI
 print('DCCA by non-linear orthogonal iterations')
-encoder_1 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
-encoder_2 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
+encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
+encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
 dcca_noi_model = dcca.DCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2], als=True)
 
 # hidden_layer_sizes are shown explicitly but these are also the defaults
@@ -325,14 +325,12 @@ dcca_noi_model.fit((train_view_1, train_view_2), epochs=epochs)
 dcca_noi_results = np.stack(
     (dcca_noi_model.train_correlations[0, 1], dcca_noi_model.predict_corr(test_view_1, test_view_2)[0, 1]))
 
-from cca_zoo import dccae
-
 # DCCAE
 print('DCCAE')
-encoder_1 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
-encoder_2 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784)
-decoder_1 = deep_models.Decoder(latent_dims=latent_dims, feature_size=784)
-decoder_2 = deep_models.Decoder(latent_dims=latent_dims, feature_size=784)
+encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
+encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
+decoder_1 = architectures.Decoder(latent_dims=latent_dims, feature_size=784)
+decoder_2 = architectures.Decoder(latent_dims=latent_dims, feature_size=784)
 dccae_model = dccae.DCCAE(latent_dims=latent_dims, encoders=[encoder_1, encoder_2], decoders=[decoder_1, decoder_2])
 
 # hidden_layer_sizes are shown explicitly but these are also the defaults
@@ -353,15 +351,14 @@ These are both implemented by the DVCCA class with private=True/False and both_e
 the encoder to the shared information Q(z_shared|x) is modelled for both x_1 and x_2 whereas if both_encoders is false
 it is modelled for x_1 as in the paper
 """
-from cca_zoo import dvcca
 
 # %%
 # DVCCA (technically bi-DVCCA)
 print('DVCCA')
-encoder_1 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
-encoder_2 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
-decoder_1 = deep_models.Decoder(latent_dims=latent_dims, feature_size=784, norm_output=True)
-decoder_2 = deep_models.Decoder(latent_dims=latent_dims, feature_size=784, norm_output=True)
+encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
+encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
+decoder_1 = architectures.Decoder(latent_dims=latent_dims, feature_size=784, norm_output=True)
+decoder_2 = architectures.Decoder(latent_dims=latent_dims, feature_size=784, norm_output=True)
 dvcca_model = dvcca.DVCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2], decoders=[decoder_1, decoder_2],
                           private=False)
 
@@ -375,12 +372,12 @@ dvcca_model_results = np.stack(
 
 # DVCCA_private (technically bi-DVCCA_private)
 print('DVCCA_private')
-encoder_1 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
-encoder_2 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
-private_encoder_1 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
-private_encoder_2 = deep_models.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
-decoder_1 = deep_models.Decoder(latent_dims=latent_dims * 2, feature_size=784, norm_output=True)
-decoder_2 = deep_models.Decoder(latent_dims=latent_dims * 2, feature_size=784, norm_output=True)
+encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
+encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
+private_encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
+private_encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784, variational=True)
+decoder_1 = architectures.Decoder(latent_dims=latent_dims * 2, feature_size=784, norm_output=True)
+decoder_2 = architectures.Decoder(latent_dims=latent_dims * 2, feature_size=784, norm_output=True)
 dvccap_model = dvcca.DVCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2], decoders=[decoder_1, decoder_2],
                            private_encoders=[private_encoder_1, private_encoder_2], private=True)
 
@@ -399,8 +396,8 @@ We can vary the encoder architecture from the default fcn to encoder/decoder bas
 """
 
 print('Convolutional DCCA')
-encoder_1 = deep_models.CNNEncoder(latent_dims=latent_dims, channels=[3, 3])
-encoder_2 = deep_models.CNNEncoder(latent_dims=latent_dims, channels=[3, 3])
+encoder_1 = architectures.CNNEncoder(latent_dims=latent_dims, channels=[3, 3])
+encoder_2 = architectures.CNNEncoder(latent_dims=latent_dims, channels=[3, 3])
 dcca_conv_model = dcca.DCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2])
 
 dcca_conv_model = deepwrapper.DeepWrapper(dcca_conv_model)
@@ -430,7 +427,7 @@ all_labels = ['linear', 'gcca', 'mcca', 'pls', 'pmd', 'elastic', 'scca', 'linear
               'gaussian kernel', 'deep CCA', 'deep generalized CCA', 'deep multiset CCA', 'deep CCAE', 'deep VCCA',
               'deep convolutional cca']
 
-from cca_zoo import plot_utils
+from cca_zoo.utils import plot_utils
 
 plot_utils.plot_results(all_results, all_labels)
 plt.show()
