+++ setup.py
@@ -6,7 +6,9 @@ with open("README.md", "r") as fh:
 with open("./requirements/basic.txt", "r") as f:
     REQUIRED_PACKAGES = f.read()
 with open("./requirements/deep.txt", "r") as f:
-    DEEP_PACKAGES = {'deep': f.read()}
+    DEEP_PACKAGES = f.read()
+with open("./requirements/probablistic.txt", "r") as f:
+    PROBABLISTIC_PACKAGES = f.read()
 
 setup(
     name='cca_zoo',
@@ -25,7 +27,8 @@ setup(
     author_email='james.chapman.19@ucl.ac.uk',
     python_requires='>=3.6',
     install_requires=REQUIRED_PACKAGES,
-    extras_require=DEEP_PACKAGES,
+    extras_require={'deep': DEEP_PACKAGES,
+                    'probablistic': PROBABLISTIC_PACKAGES},
     test_suite='tests',
     tests_require=[],
 )
