+++ tests/test_alsinnerloop.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 import numpy as np
 
-import cca_zoo.innerloop
+import cca_zoo.models.innerloop
 
 
 class TestAlsInnerLoop(TestCase):
@@ -16,8 +16,8 @@ class TestAlsInnerLoop(TestCase):
         pass
 
     def test_iterate(self):
-        pls = cca_zoo.innerloop._InnerLoop().fit(self.X, self.Y)
-        generalized_pls = cca_zoo.innerloop._InnerLoop().fit(self.X, self.Y, self.Z)
+        pls = cca_zoo.models.innerloop._InnerLoop().fit(self.X, self.Y)
+        generalized_pls = cca_zoo.models.innerloop._InnerLoop().fit(self.X, self.Y, self.Z)
         self.assertIsNone(
             np.testing.assert_almost_equal(np.linalg.norm(pls.weights[0]), np.linalg.norm(pls.weights[1]), 1))
         self.assertIsNone(
@@ -26,8 +26,9 @@ class TestAlsInnerLoop(TestCase):
                                            np.linalg.norm(generalized_pls.weights[2]), 1))
 
     def test_regularized(self):
-        park = cca_zoo.innerloop.ParkhomenkoInnerLoop(c=[0.0001, 0.0001]).fit(self.X, self.Y)
-        park_gen = cca_zoo.innerloop.ParkhomenkoInnerLoop(c=[0.0001, 0.0001], generalized=True).fit(self.X, self.Y)
+        park = cca_zoo.models.innerloop.ParkhomenkoInnerLoop(c=[0.0001, 0.0001]).fit(self.X, self.Y)
+        park_gen = cca_zoo.models.innerloop.ParkhomenkoInnerLoop(c=[0.0001, 0.0001], generalized=True).fit(self.X,
+                                                                                                           self.Y)
         params = {'c': [2, 2]}
-        pmd = cca_zoo.innerloop.PMDInnerLoop(c=[2, 2]).fit(self.X, self.Y)
-        pmd_gen = cca_zoo.innerloop.PMDInnerLoop(c=[2, 2], generalized=True).fit(self.X, self.Y)
+        pmd = cca_zoo.models.innerloop.PMDInnerLoop(c=[2, 2]).fit(self.X, self.Y)
+        pmd_gen = cca_zoo.models.innerloop.PMDInnerLoop(c=[2, 2], generalized=True).fit(self.X, self.Y)
