+++ tests/test_wrapper.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 import numpy as np
 
-import cca_zoo.wrappers
+from cca_zoo.models.wrappers import CCA, PLS, CCA_ALS, SCCA, PMD, ElasticCCA, rCCA, KCCA, KTCCA, MCCA, GCCA, TCCA
 
 np.random.seed(123)
 
@@ -20,11 +20,11 @@ class TestWrapper(TestCase):
 
     def test_unregularized_methods(self):
         latent_dims = 1
-        wrap_cca = cca_zoo.wrappers.CCA(latent_dims=latent_dims).fit(self.X, self.Y)
-        wrap_iter = cca_zoo.wrappers.CCA_ALS(latent_dims=latent_dims).fit(self.X, self.Y)
-        wrap_gcca = cca_zoo.wrappers.GCCA(latent_dims=latent_dims).fit(self.X, self.Y)
-        wrap_mcca = cca_zoo.wrappers.MCCA(latent_dims=latent_dims).fit(self.X, self.Y)
-        wrap_kcca = cca_zoo.wrappers.KCCA(latent_dims=latent_dims).fit(self.X, self.Y)
+        wrap_cca = CCA(latent_dims=latent_dims).fit(self.X, self.Y)
+        wrap_iter = CCA_ALS(latent_dims=latent_dims).fit(self.X, self.Y)
+        wrap_gcca = GCCA(latent_dims=latent_dims).fit(self.X, self.Y)
+        wrap_mcca = MCCA(latent_dims=latent_dims).fit(self.X, self.Y)
+        wrap_kcca = KCCA(latent_dims=latent_dims).fit(self.X, self.Y)
         corr_cca = wrap_cca.train_correlations[0, 1]
         corr_iter = wrap_iter.train_correlations[0, 1]
         corr_gcca = wrap_gcca.train_correlations[0, 1]
@@ -43,12 +43,12 @@ class TestWrapper(TestCase):
 
     def test_unregularized_multi(self):
         latent_dims = 5
-        wrap_cca = cca_zoo.wrappers.rCCA(latent_dims=latent_dims).fit(self.X, self.Y, self.Z)
-        wrap_iter = cca_zoo.wrappers.CCA_ALS(latent_dims=latent_dims, stochastic=False, tol=1e-12).fit(self.X, self.Y,
-                                                                                                       self.Z)
-        wrap_gcca = cca_zoo.wrappers.GCCA(latent_dims=latent_dims).fit(self.X, self.Y, self.Z)
-        wrap_mcca = cca_zoo.wrappers.MCCA(latent_dims=latent_dims).fit(self.X, self.Y, self.Z)
-        wrap_kcca = cca_zoo.wrappers.KCCA(latent_dims=latent_dims).fit(self.X, self.Y, self.Z)
+        wrap_cca = rCCA(latent_dims=latent_dims).fit(self.X, self.Y, self.Z)
+        wrap_iter = CCA_ALS(latent_dims=latent_dims, stochastic=False, tol=1e-12).fit(self.X, self.Y,
+                                                                                      self.Z)
+        wrap_gcca = GCCA(latent_dims=latent_dims).fit(self.X, self.Y, self.Z)
+        wrap_mcca = MCCA(latent_dims=latent_dims).fit(self.X, self.Y, self.Z)
+        wrap_kcca = KCCA(latent_dims=latent_dims).fit(self.X, self.Y, self.Z)
         corr_cca = wrap_cca.train_correlations[:, :, 0]
         corr_iter = wrap_iter.train_correlations[:, :, 0]
         corr_gcca = wrap_gcca.train_correlations[:, :, 0]
@@ -69,12 +69,12 @@ class TestWrapper(TestCase):
         # Test that linear regularized methods match PLS solution when using maximum regularisation
         latent_dims = 5
         c = 1
-        wrap_kernel = cca_zoo.wrappers.KCCA(latent_dims=latent_dims, c=[c, c], kernel=['linear', 'linear']).fit(self.X,
-                                                                                                                self.Y)
-        wrap_pls = cca_zoo.wrappers.PLS(latent_dims=latent_dims).fit(self.X, self.Y)
-        wrap_gcca = cca_zoo.wrappers.GCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y)
-        wrap_mcca = cca_zoo.wrappers.MCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y)
-        wrap_rCCA = cca_zoo.wrappers.rCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y)
+        wrap_kernel = KCCA(latent_dims=latent_dims, c=[c, c], kernel=['linear', 'linear']).fit(self.X,
+                                                                                               self.Y)
+        wrap_pls = PLS(latent_dims=latent_dims).fit(self.X, self.Y)
+        wrap_gcca = GCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y)
+        wrap_mcca = MCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y)
+        wrap_rCCA = rCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y)
         corr_gcca = wrap_gcca.train_correlations[0, 1]
         corr_mcca = wrap_mcca.train_correlations[0, 1]
         corr_kernel = wrap_kernel.train_correlations[0, 1]
@@ -92,18 +92,18 @@ class TestWrapper(TestCase):
         c1 = [1, 3]
         c2 = [1, 3]
         param_candidates = {'c': list(itertools.product(c1, c2))}
-        wrap_pmd = cca_zoo.wrappers.PMD(latent_dims=latent_dims).gridsearch_fit(self.X, self.Y,
-                                                                                param_candidates=param_candidates,
-                                                                                verbose=True, plot=True)
+        wrap_pmd = PMD(latent_dims=latent_dims).gridsearch_fit(self.X, self.Y,
+                                                               param_candidates=param_candidates,
+                                                               verbose=True, plot=True)
         c1 = [1e-4, 1e-5]
         c2 = [1e-4, 1e-5]
         param_candidates = {'c': list(itertools.product(c1, c2))}
-        wrap_scca = cca_zoo.wrappers.SCCA(latent_dims=latent_dims).gridsearch_fit(self.X, self.Y,
-                                                                                  param_candidates=param_candidates,
-                                                                                  verbose=True)
-        wrap_elastic = cca_zoo.wrappers.ElasticCCA(latent_dims=latent_dims).gridsearch_fit(self.X, self.Y,
-                                                                                           param_candidates=param_candidates,
-                                                                                           verbose=True)
+        wrap_scca = SCCA(latent_dims=latent_dims).gridsearch_fit(self.X, self.Y,
+                                                                 param_candidates=param_candidates,
+                                                                 verbose=True)
+        wrap_elastic = ElasticCCA(latent_dims=latent_dims).gridsearch_fit(self.X, self.Y,
+                                                                          param_candidates=param_candidates,
+                                                                          verbose=True)
         corr_pmd = wrap_pmd.train_correlations[0, 1]
         corr_scca = wrap_scca.train_correlations[0, 1]
         corr_elastic = wrap_elastic.train_correlations[0, 1]
@@ -112,21 +112,21 @@ class TestWrapper(TestCase):
         # Test that linear regularized methods match PLS solution when using maximum regularisation
         latent_dims = 5
         c = 0
-        wrap_unweighted_gcca = cca_zoo.wrappers.GCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y)
-        wrap_deweighted_gcca = cca_zoo.wrappers.GCCA(latent_dims=latent_dims, c=[c, c], view_weights=[0.5, 0.5]).fit(
+        wrap_unweighted_gcca = GCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y)
+        wrap_deweighted_gcca = GCCA(latent_dims=latent_dims, c=[c, c], view_weights=[0.5, 0.5]).fit(
             self.X, self.Y)
         corr_unweighted_gcca = wrap_unweighted_gcca.train_correlations[0, 1]
         corr_deweighted_gcca = wrap_deweighted_gcca.train_correlations[0, 1]
         # Check the correlations from each unregularized method are the same
         K = np.ones((2, self.X.shape[0]))
         K[0, 200:] = 0
-        wrap_unobserved_gcca = cca_zoo.wrappers.GCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y, K=K)
+        wrap_unobserved_gcca = GCCA(latent_dims=latent_dims, c=[c, c]).fit(self.X, self.Y, K=K)
         self.assertIsNone(np.testing.assert_array_almost_equal(corr_unweighted_gcca, corr_deweighted_gcca, decimal=1))
 
     def test_TCCA(self):
         latent_dims = 1
-        wrap_tcca = cca_zoo.wrappers.TCCA(latent_dims=latent_dims, c=[0.2, 0.2]).fit(self.X, self.Y)
-        wrap_ktcca = cca_zoo.wrappers.KTCCA(latent_dims=latent_dims, c=[0.2, 0.2]).fit(self.X, self.Y)
+        wrap_tcca = TCCA(latent_dims=latent_dims, c=[0.2, 0.2]).fit(self.X, self.Y)
+        wrap_ktcca = KTCCA(latent_dims=latent_dims, c=[0.2, 0.2]).fit(self.X, self.Y)
         corr_tcca = wrap_tcca.train_correlations[0, 1]
         corr_ktcca = wrap_ktcca.train_correlations[0, 1]
         self.assertIsNone(np.testing.assert_array_almost_equal(corr_tcca, corr_ktcca, decimal=1))
@@ -136,11 +136,11 @@ class TestWrapper(TestCase):
         c1 = [0.1, 0.2]
         c2 = [0.1, 0.2]
         param_candidates = {'c': list(itertools.product(c1, c2))}
-        wrap_unweighted_gcca = cca_zoo.wrappers.GCCA(latent_dims=latent_dims).gridsearch_fit(self.X, self.Y, folds=2,
-                                                                                             param_candidates=param_candidates)
-        wrap_deweighted_gcca = cca_zoo.wrappers.GCCA(latent_dims=latent_dims, view_weights=[0.5, 0.5]).gridsearch_fit(
+        wrap_unweighted_gcca = GCCA(latent_dims=latent_dims).gridsearch_fit(self.X, self.Y, folds=2,
+                                                                            param_candidates=param_candidates)
+        wrap_deweighted_gcca = GCCA(latent_dims=latent_dims, view_weights=[0.5, 0.5]).gridsearch_fit(
             self.X, self.Y, folds=2, param_candidates=param_candidates)
-        wrap_mcca = cca_zoo.wrappers.MCCA(latent_dims=latent_dims).gridsearch_fit(
+        wrap_mcca = MCCA(latent_dims=latent_dims).gridsearch_fit(
             self.X, self.Y, folds=2, param_candidates=param_candidates)
 
     def test_methods(self):
